Fixed bugs when downloading without titles json
1.) After switching to downloading the newest (by date) video first,
the naming was out of sync when a titles json file wasn't specified.
That is, if there were N videos then downloading the Nth video would
result in calling it "Lecture 1", the N-1 called "Lecture 2", etc.

2.) There was a regression in filenaming when not using a titles json
file such that instead of including "Lecutre N", it was just an empty
string